# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Gin personal blog backend API.",
        "title": "MyBlog API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/login": {
            "post": {
                "description": "管理员使用用户名和密码登录，成功后返回 JWT token。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "管理员登录请求体。username 示例：admin；password 示例：123456",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.AdminLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"msg\\\":\\\"登录成功\\\",\\\"data\\\":{\\\"token\\\":\\\"xxx\\\",\\\"username\\\":\\\"admin\\\",\\\"userID\\\":1}}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"参数格式错误\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"Token生成失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/admin/password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "已登录管理员修改密码。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "修改管理员密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "修改密码请求体。old_password 示例：123456；new_password 示例：654321",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"msg\\\":\\\"密码修改成功，请重新登录\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"参数校验失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "示例：{\\\"code\\\":401,\\\"msg\\\":\\\"旧密码错误\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/admin/register": {
            "post": {
                "description": "注册管理员账号，当前路由未加鉴权。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员注册",
                "parameters": [
                    {
                        "description": "管理员注册请求体。username 示例：admin；password 示例：123456",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.AdminRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"msg\\\":\\\"Welcome!\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"参数校验失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "409": {
                        "description": "示例：{\\\"code\\\":409,\\\"msg\\\":\\\"用户名已被注册\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/admin/username": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "已登录管理员修改用户名。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "修改管理员用户名",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "修改用户名请求体。old_password 示例：123456；new_username 示例：newadmin",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateUsernameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"msg\\\":\\\"用户名修改成功，请重新登录\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"参数校验失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "示例：{\\\"code\\\":401,\\\"msg\\\":\\\"用户未登录\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.AdminLoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "controllers.AdminRegisterRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6,
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2,
                    "example": "admin"
                }
            }
        },
...
        "controllers.UpdatePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6,
                    "example": "654321"
                },
                "old_password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
...
        "controllers.UpdateUsernameRequest": {
            "type": "object",
            "required": [
                "new_username",
                "old_password"
            ],
            "properties": {
                "new_username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2,
                    "example": "newadmin"
                },
                "old_password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}